Skip suspects that have no associated unblamed hunks
Pass blame to parent in `process_change`. `git`’s algorithm only seems
to keep the current suspect for unblamed hunks that were the direct
result of splitting an existing unblamed hunk because it matched with a
change. All other hunks appear to be blamed on the parent without
further checks.